Cleanup forcefield code with the new kernels.
